+++ examples/single_endpoint.js
@@ -1,4 +1,5 @@
-require('dotenv').config();
+require('dotenv')
+  .config();
 
 const debug = require('debug');
 const EnergyKit = require('../src');